// src/app/components/webApp/enum/notificationType.enum.ts
export enum NotificationTypeEnum {

   Notification_Type_NewSuggestion = 4,
   Notification_Type_RemoveSuggestion = 5,
   Notification_Type_PulseStateChanged = 8,
   Notification_Type_MsgForPulse = 9,
   Notification_Type_RequestExpire = 12,
   Notification_Type_RequestClosed = 13,
   Notification_Type_RequestWait4Payment = 14,
   Notification_Type_StopReceiveSuggestion = 16,
   Notification_Type_SuggestionChanged = 17,
   Notification_Type_InvoiceChanged = 18
}